Split tests into separate files for readability
