+++ backend-api/src/functions/common/jwks/JwksCache/tests/getSingletonInstance.test.ts
@@ -1,0 +1,9 @@
+import { InMemoryJwksCache } from "../JwksCache";
+
+describe("getSingletonInstance", () => {
+  it("Returns same instance on repeated calls", () => {
+    const first = InMemoryJwksCache.getSingletonInstance();
+    const second = InMemoryJwksCache.getSingletonInstance();
+    expect(first).toBe(second);
+  });
+});
